Export file functions from commons to their own module
