+++ cves/cve_2017_5715.py
@@ -2,7 +2,7 @@
 Support for graphviz and other modules which written for avoiding repetitive code.
 """
 import graphviz
-from modules import status, commons, constants, os_release
+from modules import status, commons, constants, file_functions, os_release
 
 
 VULNERABILITY = 'CVE-2017-5715'
@@ -29,7 +29,7 @@ MITIGATION = ''
 def check_cmdline_disabled(mitigation, debug, container_name):
     """This function checks if the spectre_v2 mitigations were not disabled by the cmdline."""
     cmdline_path = '/proc/cmdline'
-    cmdline_content = commons.file_content(cmdline_path, debug, container_name)
+    cmdline_content = file_functions.file_content(cmdline_path, debug, container_name)
     if not cmdline_content:
         print(constants.FULL_EXPLANATION_MESSAGE.format(f'The {cmdline_path} file does not exist'))
         return constants.UNSUPPORTED
@@ -48,11 +48,11 @@ def check_mitigations_components(mitigation, debug, container_name):
     """This function checks if there is a mitigation exists."""
     file_name = f'{mitigation}_enabled'
     mitigation_path = f'/sys/kernel/debug/x86/{file_name}'
-    mitigation_file = commons.check_file_existence(mitigation_path, debug, container_name)
+    mitigation_file = file_functions.check_file_existence(mitigation_path, debug, container_name)
     if not mitigation_file:
         return False
     dmesg_path = '/var/log/dmesg'
-    dmesg_content = commons.file_content(dmesg_path, debug, container_name)
+    dmesg_content = file_functions.file_content(dmesg_path, debug, container_name)
     if dmesg_content:
         print(constants.FULL_QUESTION_MESSAGE.format(f'Does {mitigation} mitigation present on the system?'))
         for line in dmesg_content:
@@ -103,7 +103,7 @@ def validate_mitigations(debug, container_name):
 def spectre_file(debug, container_name):
     """This function checks if the meltdown file contains the 'vulnerable' string in it."""
     spectre_path = '/sys/devices/system/cpu/vulnerabilities/spectre_v2'
-    spectre_content = commons.file_content(spectre_path, debug, container_name)
+    spectre_content = file_functions.file_content(spectre_path, debug, container_name)
     if not spectre_content:
         print(constants.FULL_EXPLANATION_MESSAGE.format(f'The {spectre_path} file does not exist'))
         return constants.UNSUPPORTED
@@ -122,7 +122,7 @@ def check_cpuinfo(spectre_path, debug, container_name):
     """This function checks if the cpuinfo flags field contains the ibpb string."""
     edge_case = False
     cpuinfo_path = '/proc/cpuinfo'
-    cpuinfo_content = commons.file_content(cpuinfo_path, debug, container_name)
+    cpuinfo_content = file_functions.file_content(cpuinfo_path, debug, container_name)
     if not cpuinfo_content:
         print(constants.FULL_EXPLANATION_MESSAGE.format(f'The {cpuinfo_path} file does not exist'))
         return constants.UNSUPPORTED
@@ -156,7 +156,7 @@ def check_edge_case(debug, container_name):
     print(constants.FULL_QUESTION_MESSAGE.format('Does the system meet the conditions of the edge case?'))
     if 'Red 5' in version or 'Red 6' in version:
         spectre_path = '/sys/devices/system/cpu/vulnerabilities/spectre_v2'
-        spectre_content = commons.file_content(spectre_path, debug, container_name)
+        spectre_content = file_functions.file_content(spectre_path, debug, container_name)
         if spectre_content:
             if 'Full retpoline' in spectre_content:
                 check_cpuinfo(spectre_path, debug, container_name)
